Refactor import in api tests
